requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python-headless
reportlab
pillow_heif
numpy